Moved block related methods to BlockService
