+++ src/main/java/com/example/petbuddybackend/service/block/BlockService.java
@@ -1,0 +1,69 @@
+package com.example.petbuddybackend.service.block;
+
+import com.example.petbuddybackend.entity.block.Block;
+import com.example.petbuddybackend.entity.block.BlockId;
+import com.example.petbuddybackend.repository.block.BlockRepository;
+import com.example.petbuddybackend.utils.exception.throweable.general.IllegalActionException;
+import com.example.petbuddybackend.utils.exception.throweable.general.NotFoundException;
+import com.example.petbuddybackend.utils.exception.throweable.user.AlreadyBlockedException;
+import com.example.petbuddybackend.utils.exception.throweable.user.BlockedException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class BlockService {
+
+    private static final String USER_CANNOT_BLOCK_HIMSELF_MESSAGE = "User cannot block himself";
+    private static final String USER_ALREADY_BLOCKED_MESSAGE = "User %s is already blocked";
+    private static final String BLOCK = "Block";
+
+    private final BlockRepository blockRepository;
+
+    public Block getBlock(String blockerUsername, String blockedUsername) {
+        return blockRepository.findById(new BlockId(blockerUsername, blockedUsername))
+                .orElseThrow(() -> NotFoundException.withFormattedMessage(BLOCK, blockedUsername));
+    }
+
+    public void blockUser(String blockerUsername, String blockedUsername) {
+        assertDoesNotBlockSelf(blockerUsername, blockedUsername);
+        assertNotAlreadyBlocked(blockerUsername, blockedUsername);
+
+        Block block = new Block(blockerUsername, blockedUsername);
+        blockRepository.save(block);
+    }
+
+    public void unblockUser(String blockerUsername, String blockedUsername) {
+        assertDoesNotBlockSelf(blockerUsername, blockedUsername);
+
+        Block block = getBlock(blockerUsername, blockedUsername);
+        blockRepository.delete(block);
+    }
+
+    public boolean isBlocked(String blockerUsername, String blockedUsername) {
+        return blockRepository.existsById(new BlockId(blockerUsername, blockedUsername));
+    }
+
+    public void assertNotBlockedByAny(String firstUsername, String secondUsername) {
+        assertNotBlocked(firstUsername, secondUsername);
+        assertNotBlocked(secondUsername, firstUsername);
+    }
+
+    private void assertNotBlocked(String firstUsername, String secondUsername) {
+        if(isBlocked(firstUsername, secondUsername)) {
+            throw new BlockedException(firstUsername, secondUsername);
+        }
+    }
+
+    private void assertNotAlreadyBlocked(String blockerUsername, String blockedUsername) {
+        if(isBlocked(blockerUsername, blockedUsername)) {
+            throw new AlreadyBlockedException(USER_ALREADY_BLOCKED_MESSAGE);
+        }
+    }
+
+    private void assertDoesNotBlockSelf(String blockerUsername, String blockedUsername) {
+        if(blockerUsername.equals(blockedUsername)) {
+            throw new IllegalActionException(USER_CANNOT_BLOCK_HIMSELF_MESSAGE);
+        }
+    }
+}
