+++ src/test/java/com/example/petbuddybackend/service/user/UserServiceTest.java
@@ -1,8 +1,6 @@
 package com.example.petbuddybackend.service.user;
 
 import com.example.petbuddybackend.dto.user.UserProfilesData;
-import com.example.petbuddybackend.entity.block.Block;
-import com.example.petbuddybackend.entity.block.BlockId;
 import com.example.petbuddybackend.entity.photo.PhotoLink;
 import com.example.petbuddybackend.entity.user.AppUser;
 import com.example.petbuddybackend.entity.user.Role;
@@ -13,10 +11,7 @@ import com.example.petbuddybackend.repository.user.ClientRepository;
 import com.example.petbuddybackend.service.photo.PhotoService;
 import com.example.petbuddybackend.testutils.mock.MockUserProvider;
 import com.example.petbuddybackend.utils.exception.throweable.InvalidRoleException;
-import com.example.petbuddybackend.utils.exception.throweable.general.IllegalActionException;
 import com.example.petbuddybackend.utils.exception.throweable.general.NotFoundException;
-import com.example.petbuddybackend.utils.exception.throweable.user.AlreadyBlockedException;
-import com.example.petbuddybackend.utils.exception.throweable.user.BlockedException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -269,139 +264,6 @@ public class UserServiceTest {
         // When & Then
         assertThrows(NotFoundException.class, () -> userService.deleteProfilePicture(EMAIL));
         verify(photoService, never()).deletePhoto(any(String.class));
-    }
-
-    @Test
-    void blockUser_whenUserBlocksAnotherUser_shouldSucceed() {
-        // Given
-        String blockerUsername = "blocker@example.com";
-        String blockedUsername = "blocked@example.com";
-
-        when(blockRepository.existsById(any(BlockId.class))).thenReturn(false); // User not already blocked
-
-        // When
-        userService.blockUser(blockerUsername, blockedUsername);
-
-        // Then
-        verify(blockRepository).save(any(Block.class));
-    }
-
-    @Test
-    void blockUser_whenUserBlocksThemselves_shouldThrowException() {
-        // Given
-        String username = "same@example.com";
-
-        // When & Then
-        IllegalActionException exception = assertThrows(IllegalActionException.class, () ->
-                userService.blockUser(username, username));
-        assertEquals("User cannot block himself", exception.getMessage());
-    }
-
-    @Test
-    void blockUser_whenUserIsAlreadyBlocked_shouldThrowException() {
-        // Given
-        String blockerUsername = "blocker@example.com";
-        String blockedUsername = "blocked@example.com";
-
-        when(blockRepository.existsById(any(BlockId.class))).thenReturn(true); // User already blocked
-
-        // When & Then
-        assertThrows(AlreadyBlockedException.class, () ->
-                userService.blockUser(blockerUsername, blockedUsername));
-    }
-
-    @Test
-    void unblockUser_whenBlockExists_shouldSucceed() {
-        // Given
-        String blockerUsername = "blocker@example.com";
-        String blockedUsername = "blocked@example.com";
-
-        when(blockRepository.findById(any(BlockId.class))).thenReturn(Optional.of(new Block(blockerUsername, blockedUsername)));
-
-        // When
-        userService.unblockUser(blockerUsername, blockedUsername);
-
-        // Then
-        verify(blockRepository).delete(any(Block.class));
-    }
-
-    @Test
-    void unblockUser_whenBlockDoesNotExist_shouldThrowException() {
-        // Given
-        String blockerUsername = "blocker@example.com";
-        String blockedUsername = "blocked@example.com";
-
-        when(blockRepository.findById(any(BlockId.class))).thenReturn(Optional.empty());
-
-        // When & Then
-        assertThrows(NotFoundException.class, () ->
-                userService.unblockUser(blockerUsername, blockedUsername));
-    }
-
-    @Test
-    void unblockUser_whenUserUnblocksThemselves_shouldThrowException() {
-        // Given
-        String username = "same@example.com";
-
-        // When & Then
-        assertThrows(IllegalActionException.class, () ->
-                userService.unblockUser(username, username));
-    }
-
-    @Test
-    void isBlocked_whenUserIsBlocked_shouldReturnTrue() {
-        // Given
-        String blockerUsername = "blocker@example.com";
-        String blockedUsername = "blocked@example.com";
-
-        when(blockRepository.existsById(any(BlockId.class))).thenReturn(true);
-
-        // When
-        boolean result = userService.isBlocked(blockerUsername, blockedUsername);
-
-        // Then
-        assertTrue(result);
-    }
-
-    @Test
-    void isBlocked_whenUserIsNotBlocked_shouldReturnFalse() {
-        // Given
-        String blockerUsername = "blocker@example.com";
-        String blockedUsername = "blocked@example.com";
-
-        when(blockRepository.existsById(any(BlockId.class))).thenReturn(false);
-
-        // When
-        boolean result = userService.isBlocked(blockerUsername, blockedUsername);
-
-        // Then
-        assertFalse(result);
-    }
-
-    @Test
-    void assertNotBlockedByAny_whenBlockedByOtherUser_shouldThrowException() {
-        // Given
-        String firstUsername = "user1@example.com";
-        String secondUsername = "user2@example.com";
-
-        when(blockRepository.existsById(eq(new BlockId(firstUsername, secondUsername)))).thenReturn(true);
-
-        // When & Then
-        assertThrows(BlockedException.class, () ->
-                userService.assertNotBlockedByAny(firstUsername, secondUsername));
-    }
-
-    @Test
-    void assertNotBlockedByAny_whenBlockedByOtherUserReversed_shouldThrowException() {
-        // Given
-        String firstUsername = "user1@example.com";
-        String secondUsername = "user2@example.com";
-
-        when(blockRepository.existsById(eq(new BlockId(secondUsername, firstUsername)))).thenReturn(true);
-
-        // When & Then
-        assertThrows(BlockedException.class, () ->
-                userService.assertNotBlockedByAny(firstUsername, secondUsername));
     }
 
     @Test
